877: Renames get-open-workload-report to get-open-workload-reports
