+++ app/services/data/get-open-workload-reports.js
