+++ test/unit/app/services/test-close-previous-workload-report.js
@@ -3,7 +3,7 @@ const proxyquire = require('proxyquire')
 const sinon = require('sinon')
 require('sinon-bluebird')
 
-var getOpenWorkloadReport
+var getOpenWorkloadReports
 var updateWorkloadEffectiveTo
 var closePreviousWorkloadReport
 
@@ -25,21 +25,21 @@ const currentWorkloads = [
 
 describe('services/task-counter', function () {
   before(function (done) {
-    getOpenWorkloadReport = sinon.stub()
+    getOpenWorkloadReports = sinon.stub()
     updateWorkloadEffectiveTo = sinon.stub()
 
     closePreviousWorkloadReport = proxyquire('../../../../app/services/close-previous-workload-report', {
-      './data/get-open-workload-report': getOpenWorkloadReport,
+      './data/get-open-workload-reports': getOpenWorkloadReports,
       './data/update-workload-report-effective-to': updateWorkloadEffectiveTo
     })
     done()
   })
 
-  it('should call get-open-workload-report and update-workload-report-effective-to', function (done) {
-    getOpenWorkloadReport.resolves(currentWorkloads)
+  it('should call get-open-workload-reports and update-workload-report-effective-to', function (done) {
+    getOpenWorkloadReports.resolves(currentWorkloads)
     updateWorkloadEffectiveTo.resolves()
     closePreviousWorkloadReport(newWorkloadReportId).then(function () {
-      expect(getOpenWorkloadReport.called).to.be.true //eslint-disable-line
+      expect(getOpenWorkloadReports.called).to.be.true //eslint-disable-line
       expect(updateWorkloadEffectiveTo.calledWith(previousWorkloadReportId, now)).to.be.true //eslint-disable-line
       done()
     })
